Added BaseController and BaseViewModel for Handling insertion Dogovors and RmDogovors RmdogovorController to be implemented...

+++ NEO SYSTEM TECHNOLOGY/Entity/Dogovor.cs
@@ -10,7 +10,7 @@ namespace NEO_SYSTEM_TECHNOLOGY.Entity
         public string OrderHeader { get; set; }
 
         
-        public decimal DogovorSum { get; set; }
+        public double DogovorSum { get; set; }
 
        
         public DateTime StartDate { get; set; }
